typo

+++ src/alchemy/y/alchemy/library.hpp
@@ -101,7 +101,7 @@ namespace upsylon
                 {
                     const Species &sp = **it;
                     lib.print(os << ' ',sp);
-                    os << vformat(" : %3ld @%u\n", sp.z, unsigned(sp.indx));
+                    os << vformat(" : %3ld @%u\n", static_cast<long>(sp.z), unsigned(sp.indx));
                 }
                 os << '}';
                 return os;
